Remove `verify` (#8)
MLIR has the concept of verify to ensure that an operation is legal after it has been modified. On first thought, this made a lot of sense for XRCF too. The reason is that objects inside the codebase are constantly mutated on the fly. It therefore would then be nice to have one `verify` step which can be used to occasionally check whether everything is still okay.

However, I'm going to try to see how far we can get without having a verifier. Unlike C++, it is much easier in Rust to inspect the operations directly. C++ is more limited to checking the textual representation. With the easier inspections, I suspect that XRCF can do all the verification inside the tests and hence can live without it inside the production part of the codebase.