Create `xr-example` (#13)
Projects like Jax, Triton, Mojo, and Core ML all take high-level
operations, like `transpose`, at the compiler level so that these
operations can be optimized. Hence, the aim of `xr-example` should also
be to compile high-level operations. Therefore, this PR replaces the
Python compiler `pythonc` by an "Example" language compiler
`xr-example`. This compiler should be able to generate library files
which can be included in other languages. That would be a way to make
higher-level ops available to other languages.

(A small note on Mojo. The examples often operate on individual values
**BUT** the types of the values can be set to SIMD types which means
that Mojo probably probably has the possibility to optimize these
operations to higher-level ops.)

This doesn't mean that xrcf will not be useful for building
"lower-level" compilers. It's just that compiling high-level operations
is currently a problem that many projects are dealing with.