Automatically rename blocks (#34)
Just like arguments and op results, block names should be optional too.
This means that operands can use pointers to blocks instead block
labels. When blocks are auto-named, the APIs can be simplified since we
only have to generate objects and can then refer to these objects during
transformations. With labels, it's much trickier to avoid accidentally
having duplicate labels (read: labels imply more state then just
objects).

This means that operands that use block labels need to be replaced by
pointers during parsing. Since I prefer to keep the parser to one pass,
the strategy is:

1. When parsing operands, just parse the block label and remember that.
2. When parsing a block, find all the operands that have the current
block label and replace the labels with pointers.

Then refresh the name **before** and **not** during printing. If set
during printing, then anything that refers to the block will show a
wrong name since the pointers occur **before** the actual block.
(Contrary to variables which are always defined before use.)

It would be possible to add a map inside a `Region` that could for each
label contain a pointer to the block. Then only one pointer has to be
updated. However, this to me seems like a layer of indirection which I
rather avoid for now.

As a fix of an earlier introduced problem, this PR also removes
`BlockDest` since it should have been a pointer to the block all along.
The pointer to the block is now a `Value::BlockPtr` which means that
they can just be stored as operands instead of special fields on ops.