Remove `Arc` helpers such as `GuardedOp` (#38)
The idea was to have special traits with convenience methods 

```rust
impl GuardedRegion for Arc<RwLock<Region>> {
    fn add_empty_block(&self) -> UnsetBlock {
        self.rd().add_empty_block()
    }
}
```
to allow writing
```rust
region.add_empty_block()
```
instead of
```rust
region.try_read().unwrap().add_empty_block()
```

This had two problems. One was that it of course takes extra code to
implement and clients would be needed to always include the trait. The
bigger problem though is that this layer of indirection hides the
docstring from the original method unless you copy the docstring.