Revert everything up to "Go via MLIR"
This reverts commit c490c66fab171323113cdf2dc47965fc7279ae5a.

Cranelift compiles wasm/wat to native not arbitrary IR to wasm/wat.

I think I'm just going to go forward with manually generating wat files.
Maybe it's not the very best for performance, but who cares at this
point. Getting something to work is way more important. On the question
whether to target wat or wasm, I think wat is fine. Also here, maybe
it's more efficient to go straight to wasm, but it's not necessarily
easier. wat is probably much easier to debug.